+++ src/main/java/tk/estecka/nokebab/mixin/PaintingEntityMixin.java
@@ -16,32 +16,47 @@ import tk.estecka.nokebab.PaintingState;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 
+@Unique
 @Mixin(PaintingEntity.class)
 public abstract class PaintingEntityMixin
 extends AbstractDecorationEntity
 implements IPaintingEntityDuck
 {
-	static private final TrackedData<String> MISSING_VARIANT = DataTracker.registerData(PaintingEntity.class, TrackedDataHandlerRegistry.STRING);
+	static private final String VARIANT_NBT_KEY = "variant";
+	static private final TrackedData<String> MISSING_TRACKER;
+
+	static {
+		if (NoKebab.areCustomTrackersEnabled())
+			MISSING_TRACKER = DataTracker.registerData(PaintingEntity.class, TrackedDataHandlerRegistry.STRING);
+		else
+			MISSING_TRACKER = null;
+	}
+	
+
+	private @NotNull String MISSING_VARIANT = "";
 
 	private PaintingEntityMixin(){ super(null, null); }
-
 	@Shadow public abstract RegistryEntry<PaintingVariant> getVariant();
 	@Shadow public abstract void setVariant(RegistryEntry<PaintingVariant> variant);
 
 
 	@Override
 	public @NotNull String	nokebab$GetMissingName(){
-		return this.getDataTracker().get(MISSING_VARIANT);
+		return NoKebab.areCustomTrackersEnabled() ? this.dataTracker.get(MISSING_TRACKER) : MISSING_VARIANT;
 	}
 	@Override
 	public void	nokebab$SetMissingName(@NotNull String value){
-		this.getDataTracker().set(MISSING_VARIANT, value);
+		if (NoKebab.areCustomTrackersEnabled())
+			this.dataTracker.set(MISSING_TRACKER, value);
+		else
+			this.MISSING_VARIANT = value;
 	}
 
 	@Override
@@ -56,8 +71,9 @@ implements IPaintingEntityDuck
 
 
 	@Inject( method="initDataTracker", at=@At("HEAD") )
-	private void	InitMissingTracker(CallbackInfo info){
-		this.getDataTracker().startTracking(MISSING_VARIANT, "");
+	private void	InitMissingTracker(DataTracker.Builder builder, CallbackInfo info){
+		if (NoKebab.areCustomTrackersEnabled())
+			builder.add(MISSING_TRACKER, "");
 	}
 
 	@Inject( method="setVariant", at=@At("HEAD") )
@@ -76,7 +92,7 @@ implements IPaintingEntityDuck
 		if (missingName.isEmpty())
 			original.call(nbt, entry);
 		else {
-			nbt.putString(PaintingEntity.VARIANT_NBT_KEY, missingName);
+			nbt.putString(VARIANT_NBT_KEY, missingName);
 			if (!entry.matchesId(Registries.PAINTING_VARIANT.getDefaultId())){
 				NoKebab.LOGGER.error("Painting is Missingno, but active variant is not the default one: {} {} ", this.getPos(), this.getUuid());
 				NoKebab.LOGGER.error("Known: \"{}\" Active: {}", missingName, entry.getKey());
@@ -86,7 +102,7 @@ implements IPaintingEntityDuck
 
 	@Inject( method="readCustomDataFromNbt", at=@At("TAIL") )
 	private void	preserveMissingVariantFromNBT(NbtCompound nbt, CallbackInfo info){
-		String nbtString = nbt.getString(PaintingEntity.VARIANT_NBT_KEY);
+		String nbtString = nbt.getString(VARIANT_NBT_KEY);
 		if (nbtString.isEmpty())
 			return;